Instance factory extracted / reset password enabled on higher envs by default
