+++ src/main/kotlin/com/cognifide/gradle/aem/common/instance/check/CheckProgress.kt
@@ -1,6 +1,7 @@
 package com.cognifide.gradle.aem.common.instance.check
 
 import com.cognifide.gradle.aem.common.instance.Instance
+import com.cognifide.gradle.common.utils.Patterns
 import org.apache.commons.lang3.time.StopWatch
 
 class CheckProgress(val instance: Instance) {
@@ -27,5 +28,23 @@ class CheckProgress(val instance: Instance) {
 
     val summary: String get() = "${instance.name}: ${currentCheck?.summary ?: "In progress"}"
 
-    val shortSummary: String get() = "${instance.name}: ${currentCheck?.status ?: "In progress"}"
+    val abbreviatedSummary: String get() {
+        val sign = currentCheck?.let { if (it.done) "+" else "-" } ?: "~"
+        val parts = mutableListOf(sign)
+
+        currentCheck?.summary?.let { summary ->
+            if (Patterns.wildcard(summary, "* (*)")) {
+                val text = summary.substringBefore(" (")
+                val number = summary.substringAfter(" (").removeSuffix(")")
+                parts.add(text.firstLetters())
+                parts.add(number)
+            } else {
+                parts.add(summary.firstLetters())
+            }
+        }
+
+        return "${instance.name}: ${parts.joinToString("|")}"
+    }
+
+    private fun String.firstLetters() = this.split(" ").mapNotNull { it.firstOrNull() }.joinToString("")
 }
